fix: dead middleware auth, publicly-cached admin/customer pages, and real calendar-ID leak
Full audit triggered by a stale /admin/settings page: found that middleware.ts's
cookie-based admin/customer gate was dead code (checked admin-token/auth-token
cookies the app never sets, gated behind ENFORCE_COOKIE_AUTH which was never
set in Vercel), and separately that 7 admin pages had 'use client' + export
const dynamic = 'force-dynamic' in the same file — Next.js silently ignores
route-segment config in client modules, so these were never actually dynamic
despite the code appearing to force it. They were being served from Next.js's
static/ISR cache with a public Cache-Control header, confirmed servable in
full to a completely unauthenticated curl request (up to 34 days stale in one
case).

Changes:
- middleware.ts: removed the dead cookie-auth functions/redirects. Real access
  control was already, and remains, entirely client-side (admin/layout.tsx's
  Firebase auth check) plus per-route API guards (requireAdmin/requireAuth/
  requireOwnerOrAdmin) — confirmed intact by an 8-angle review. Replaced the
  blanket public cache-control with a private/no-store treatment for admin +
  customer-authenticated paths, keeping short public caching only for
  genuinely public marketing/legal pages.
- Split the 7 broken 'use client'+dynamic pages (bookings, costs, health,
  messages, messages/[threadId], pricing, settings) into a thin Server
  Component page.tsx + a new *Client.tsx, matching the pattern /book and
  /dashboard already used correctly.
- Added export const dynamic = 'force-dynamic' to ~15 other admin/customer
  pages that previously used ISR (revalidate) or got silently static-optimized.
- AdminGoogleCalendarView was a Server Component reading
  NEXT_PUBLIC_GOOGLE_CALENDAR_* env vars directly and baking the real
  calendar ID/embed URL into server-rendered HTML with zero auth check —
  moved that into a new requireAdmin-gated route (/api/admin/calendar-embed)
  and converted the component to fetch it client-side.
- Added scripts/eslint-rules/fairfield-custom-rules.js:
  no-client-route-segment-config, erroring on 'use client' + dynamic/
  revalidate/fetchCache/runtime in the same file, so this exact bug class
  can't silently reappear.

Review findings addressed: the new calendar-embed route now checks res.ok
before parsing (was rendering a misleading "not configured" message on auth
failures), has a try/catch matching sibling admin routes, and sets
Cache-Control: private, max-age=300 so it doesn't re-run Firebase Admin +
Firestore auth checks on every single page visit for data that only changes
on redeploy.

Known, accepted gap (not closed by this PR): there is still no real
server-side access control for /admin/* — an unauthenticated request gets a
genuine fresh response, just no longer a shared stale one. Closing that
would mean wiring middleware to decode the real booking_session cookie
server-side, a larger change intentionally out of scope here.
