+++ src/app/(admin)/admin/schedules/calendar/page.tsx
@@ -1,5 +1,7 @@
 import { AdminGoogleCalendarView } from '../_components/AdminGoogleCalendarView';
 
+export const dynamic = 'force-dynamic';
+
 export default function DriverCalendarPage() {
   return <AdminGoogleCalendarView />;
 }
