+++ src/app/(customer)/feedback/[id]/page.tsx
@@ -7,7 +7,7 @@ import {
 import FeedbackClient from './FeedbackClient';
 
 // Enable ISR for dynamic content updates
-export const revalidate = 3600; // Revalidate every hour
+export const dynamic = 'force-dynamic';
 
 // Main page component (Server Component)
 export default async function FeedbackPage({ params }: { params: Promise<{ id: string }> }) {
