+++ src/app/(customer)/payment-success/page.tsx
@@ -7,7 +7,7 @@ import {
 import PaymentSuccessClient from './PaymentSuccessClient';
 
 // Enable ISR for dynamic content updates
-export const revalidate = 3600; // Revalidate every hour
+export const dynamic = 'force-dynamic';
 
 // Main page component (Server Component)
 export default async function PaymentSuccessPage() {
