+++ src/app/(customer)/payments/page.tsx
@@ -7,7 +7,7 @@ import {
 import CustomerPaymentsClient from './CustomerPaymentsClient';
 
 // Enable ISR for dynamic content updates
-export const revalidate = 3600; // Revalidate every hour
+export const dynamic = 'force-dynamic';
 
 // Main page component (Server Component)
 export default function CustomerPaymentsPage() {
